Fix security issue

+++ files/views.py
@@ -6,6 +6,7 @@ from django.http import HttpResponseRedirect, HttpResponse
 from scienceshaped import admin_history
 import os
 from django.conf import settings
+from authentication import templatetags
 
 def findId(title):
     number = 1
@@ -47,8 +48,8 @@ def renameImage(instance, title):
     instance.save()
     return instance
 
-    if request.method == 'POST':
 def imageUpload(request):
+    if request.method == 'POST' and templatetags.tags.inGroup(request.user, 'editor'):
         form = ImageUpload(request.POST, request.FILES)
         if form.is_valid():
             img = saveImage(request.FILES['file'], str(form.cleaned_data['title']), str(form.cleaned_data['tags']))
@@ -93,16 +94,17 @@ def images(request):
     return render(request, 'images.html', context)
 
 def imageDelete(request, image_id):
-    try:
-        image = Image.objects.get(pk=image_id)
-        image.delete()
-    except Image.DoesNotExist:
-        pass
+    if templatetags.tags.inGroup(request.user, 'editor'):
+        try:
+            image = Image.objects.get(pk=image_id)
+            image.delete()
+        except Image.DoesNotExist:
+            pass
 
     return HttpResponseRedirect('/files/images')
 
 def imageEdit(request, image_id):
-    if request.method == 'POST':
+    if request.method == 'POST' and templatetags.tags.inGroup(request.user, 'editor'):
         form = ImageEdit(request.POST)
         if form.is_valid():
             try:
